refactor(ui): simplify hydration with automatic strategy [#703]
Replace 6 manual hydration strategies (eager, lazy, visible, idle,
interaction, media) with a single autoStrategy that uses
IntersectionObserver with 200px rootMargin. Above-fold components
hydrate immediately; below-fold components hydrate when scrolled near.

- Remove public exports: eagerStrategy, lazyStrategy, visibleStrategy,
  interactionStrategy, idleStrategy, mediaStrategy
- Remove unused registry field from MountOptions
- Update docs and cheat sheet

Closes #703
