+++ packages/ui/.api-cheat-sheet.md
@@ -51,8 +51,7 @@
 - `useSearchParams(searchParams)` - Use search params
 
 ### Hydration
-- `hydrate(registry, strategy?)` - Hydrate components
-- Strategies: `eagerStrategy`, `idleStrategy`, `visibleStrategy`, `interactionStrategy`, `lazyStrategy`, `mediaStrategy`
+- `hydrate(registry)` - Hydrate components (automatic strategy via IntersectionObserver)
 
 ## Signal API Details
