+++ packages/ui/src/hydrate/__tests__/hydrate.test.ts
@@ -43,10 +43,9 @@ describe('hydrate()', () => {
     globalThis.IntersectionObserver = origIntersectionObserver;
   });
 
-  // IT-5B-1: Hydration bootstraps interactive components from server HTML
   it('bootstraps interactive component from server-rendered HTML', async () => {
     document.body.innerHTML = `
-      <div data-v-id="Counter" data-v-key="c1" hydrate="eager">
+      <div data-v-id="Counter" data-v-key="c1">
         <script type="application/json">{"initial":0}</script>
         <button>0</button>
       </div>
@@ -58,7 +57,6 @@ describe('hydrate()', () => {
     const CounterComponent: ComponentFunction = (props, el) => {
       receivedProps = props;
       receivedEl = el;
-      // Simulate hydration: bind click handler
       const button = el.querySelector('button');
       if (!button) return;
       let count = (props.initial as number) ?? 0;
@@ -74,7 +72,13 @@ describe('hydrate()', () => {
 
     hydrate(registry);
 
-    // Wait for async resolution
+    // Trigger auto-strategy IO callback
+    const el = document.querySelector('[data-v-id]');
+    observeCallback(
+      [{ isIntersecting: true, target: el } as IntersectionObserverEntry],
+      {} as IntersectionObserver,
+    );
+
     await waitFor(() => {
       expect(receivedEl).not.toBeNull();
     });
@@ -87,10 +91,9 @@ describe('hydrate()', () => {
     expect(button?.textContent).toBe('1');
   });
 
-  // IT-5B-2: Lazy hydration uses IntersectionObserver
-  it('lazy hydration delays until element is visible', () => {
+  it('auto strategy delays hydration until element is near viewport', () => {
     document.body.innerHTML = `
-      <div data-v-id="LazyComponent" data-v-key="l1" hydrate="lazy">
+      <div data-v-id="LazyComponent" data-v-key="l1">
         <script type="application/json">{}</script>
         <div>Content</div>
       </div>
@@ -120,36 +123,9 @@ describe('hydrate()', () => {
     expect(hydrateSpy).toHaveBeenCalled();
   });
 
-  // IT-5B-3: Interaction hydration triggers on first user event
-  it('interaction hydration triggers on first click', () => {
-    document.body.innerHTML = `
-      <div data-v-id="InteractiveComponent" data-v-key="i1" hydrate="interaction">
-        <script type="application/json">{}</script>
-        <button>Click me</button>
-      </div>
-    `;
-
-    const hydrateSpy = vi.fn();
-    const registry: ComponentRegistry = {
-      InteractiveComponent: () => {
-        hydrateSpy();
-        return Promise.resolve({
-          default: () => {},
-        });
-      },
-    };
-
-    hydrate(registry);
-    expect(hydrateSpy).not.toHaveBeenCalled();
-
-    // Click the button -- event bubbles to the container
-    document.querySelector('button')?.click();
-    expect(hydrateSpy).toHaveBeenCalled();
-  });
-
   it('reports error via console.error when chunk load fails during hydration', async () => {
     document.body.innerHTML = `
-      <div data-v-id="BrokenComponent" data-v-key="b1" hydrate="eager">
+      <div data-v-id="BrokenComponent" data-v-key="b1">
         <script type="application/json">{}</script>
         <div>Content</div>
       </div>
@@ -164,6 +140,13 @@ describe('hydrate()', () => {
 
     hydrate(registry);
 
+    // Trigger auto-strategy IO callback
+    const el = document.querySelector('[data-v-id]');
+    observeCallback(
+      [{ isIntersecting: true, target: el } as IntersectionObserverEntry],
+      {} as IntersectionObserver,
+    );
+
     await waitFor(() => {
       expect(consoleSpy).toHaveBeenCalledWith(
         '[hydrate] Failed to hydrate component "BrokenComponent":',
@@ -174,8 +157,7 @@ describe('hydrate()', () => {
     consoleSpy.mockRestore();
   });
 
-  // Default strategy is lazy
-  it('defaults to lazy strategy when no hydrate attribute', () => {
+  it('uses auto strategy when no hydrate attribute is present', () => {
     document.body.innerHTML = `
       <div data-v-id="DefaultComponent" data-v-key="d1">
         <script type="application/json">{}</script>
@@ -194,13 +176,13 @@ describe('hydrate()', () => {
     };
 
     hydrate(registry);
-    // Should not be called yet (lazy waits for intersection)
+    // Auto strategy waits for intersection, should not fire immediately
     expect(hydrateSpy).not.toHaveBeenCalled();
   });
 
   it('sets data-v-hydrated attribute after hydration', async () => {
     document.body.innerHTML = `
-      <div data-v-id="MarkedComponent" data-v-key="m1" hydrate="eager">
+      <div data-v-id="MarkedComponent" data-v-key="m1">
         <script type="application/json">{}</script>
         <div>Content</div>
       </div>
@@ -216,17 +198,23 @@ describe('hydrate()', () => {
 
     hydrate(registry);
 
+    // Trigger auto-strategy IO callback
+    const el = document.querySelector('[data-v-id="MarkedComponent"]');
+    observeCallback(
+      [{ isIntersecting: true, target: el } as IntersectionObserverEntry],
+      {} as IntersectionObserver,
+    );
+
     await waitFor(() => {
       expect(componentFn).toHaveBeenCalledOnce();
     });
 
-    const el = document.querySelector('[data-v-id="MarkedComponent"]');
     expect(el?.getAttribute('data-v-hydrated')).toBe('');
   });
 
   it('does not double-hydrate a component when hydrate() is called twice', async () => {
     document.body.innerHTML = `
-      <div data-v-id="DoubleComponent" data-v-key="dbl1" hydrate="eager">
+      <div data-v-id="DoubleComponent" data-v-key="dbl1">
         <script type="application/json">{}</script>
         <div>Content</div>
       </div>
@@ -242,6 +230,13 @@ describe('hydrate()', () => {
 
     hydrate(registry);
 
+    // Trigger auto-strategy IO callback
+    const el = document.querySelector('[data-v-id]');
+    observeCallback(
+      [{ isIntersecting: true, target: el } as IntersectionObserverEntry],
+      {} as IntersectionObserver,
+    );
+
     await waitFor(() => {
       expect(componentFn).toHaveBeenCalledOnce();
     });
@@ -256,7 +251,7 @@ describe('hydrate()', () => {
 
   it('skips elements that already have data-v-hydrated attribute', () => {
     document.body.innerHTML = `
-      <div data-v-id="AlreadyHydrated" data-v-key="ah1" hydrate="eager" data-v-hydrated>
+      <div data-v-id="AlreadyHydrated" data-v-key="ah1" data-v-hydrated>
         <script type="application/json">{}</script>
         <div>Content</div>
       </div>
@@ -274,91 +269,5 @@ describe('hydrate()', () => {
 
     // Component should NOT be hydrated because it already has the marker
     expect(componentFn).not.toHaveBeenCalled();
-  });
-
-  it('routes idle strategy to idleStrategy', () => {
-    const origRequestIdleCallback = globalThis.requestIdleCallback;
-    globalThis.requestIdleCallback = vi.fn((cb: IdleRequestCallback) => {
-      cb({} as IdleDeadline);
-      return 1;
-    }) as typeof globalThis.requestIdleCallback;
-
-    document.body.innerHTML = `
-      <div data-v-id="IdleComponent" data-v-key="idle1" hydrate="idle">
-        <script type="application/json">{}</script>
-        <div>Content</div>
-      </div>
-    `;
-
-    const hydrateSpy = vi.fn();
-    const registry: ComponentRegistry = {
-      IdleComponent: () => {
-        hydrateSpy();
-        return Promise.resolve({ default: () => {} });
-      },
-    };
-
-    hydrate(registry);
-    expect(hydrateSpy).toHaveBeenCalled();
-
-    globalThis.requestIdleCallback = origRequestIdleCallback;
-  });
-
-  it('routes visible strategy to visibleStrategy', () => {
-    document.body.innerHTML = `
-      <div data-v-id="VisibleComponent" data-v-key="vis1" hydrate="visible">
-        <script type="application/json">{}</script>
-        <div>Content</div>
-      </div>
-    `;
-
-    const hydrateSpy = vi.fn();
-    const registry: ComponentRegistry = {
-      VisibleComponent: () => {
-        hydrateSpy();
-        return Promise.resolve({ default: () => {} });
-      },
-    };
-
-    hydrate(registry);
-    // visible strategy uses IntersectionObserver, so should not fire immediately
-    expect(hydrateSpy).not.toHaveBeenCalled();
-  });
-
-  it('routes media strategy with query attribute', () => {
-    let mediaChangeHandler: ((event: { matches: boolean }) => void) | undefined;
-    const origMatchMedia = globalThis.matchMedia;
-    globalThis.matchMedia = ((query: string) => ({
-      matches: false,
-      media: query,
-      addEventListener: (_event: string, handler: (event: { matches: boolean }) => void) => {
-        mediaChangeHandler = handler;
-      },
-      removeEventListener: vi.fn(),
-    })) as typeof globalThis.matchMedia;
-
-    document.body.innerHTML = `
-      <div data-v-id="MediaComponent" data-v-key="med1" hydrate="media" hydrate-media="(min-width: 768px)">
-        <script type="application/json">{}</script>
-        <div>Content</div>
-      </div>
-    `;
-
-    const hydrateSpy = vi.fn();
-    const registry: ComponentRegistry = {
-      MediaComponent: () => {
-        hydrateSpy();
-        return Promise.resolve({ default: () => {} });
-      },
-    };
-
-    hydrate(registry);
-    expect(hydrateSpy).not.toHaveBeenCalled();
-
-    // Simulate media query match
-    mediaChangeHandler?.({ matches: true });
-    expect(hydrateSpy).toHaveBeenCalled();
-
-    globalThis.matchMedia = origMatchMedia;
   });
 });
