+++ packages/ui/src/hydrate/index.ts
@@ -2,11 +2,3 @@ export type { ComponentFunction, ComponentLoader, ComponentRegistry } from './co
 export { resolveComponent } from './component-registry';
 export { hydrate } from './hydrate';
 export { deserializeProps } from './props-deserializer';
-export {
-  eagerStrategy,
-  idleStrategy,
-  interactionStrategy,
-  lazyStrategy,
-  mediaStrategy,
-  visibleStrategy,
-} from './strategies';
